+++ docs/CHANGELOG.md
@@ -32,6 +32,7 @@ The entire project (i.e., the GraphQL API, REST API, and server) uses the same v
 - Paginate `Query.searchMessages`.
 - Paginate `Query.readChats`.
 - Paginate `Query.searchChats`.
+- Paginate `Query.searchPublicChats`.
 
 ### Removed